refactor: make separation of responsabilities between auth and user controller
